+++ CHANGELOG.md
@@ -30,11 +30,12 @@ The present file will list all changes made to the project; according to the
   The previous recipient option still exists if needed. This replacement will only happen once during the upgrade.
 
 #### Deprecated
-- Usage of "users_id_validate" parameter in "ajax/dropdownValidator.php" and "front/commonitilvalidation.form.php".
+- Usage of `users_id_validate` parameter in `ajax/dropdownValidator.php` and `front/commonitilvalidation.form.php`.
 - `front/change_problem.form.php` script usage.
 - `front/change_ticket.form.php` script usage.
 - `front/problem_ticket.form.php` script usage.
 - `front/ticket_ticket.form.php` script usage.
+- Usage of `users_id_validate` field in `CommonITILValidation`.
 - `Ticket` `link_to_problem` massive action is deprecated. Use `CommonITILObject_CommonITILObject` `add` massive action instead.
 - `Ticket_Ticket` `add` massive action is deprecated. Use `CommonITILObject_CommonITILObject` `add` massive action instead.
 - `Ticket_Ticket::checkParentSon()`